+++ dila/data/__init__.py
@@ -11,9 +11,12 @@ from dila.data.languages import get_language
 from dila.data.languages import get_languages
 from dila.data.engine import setup
 from dila.data.engine import shutdown_session
+from dila.data.po_metadata import get_po_metadata
+from dila.data.po_metadata import update_po_metadata
 
 
 __all__ = [
     add_or_update_base_string, delete_old_strings, get_translated_string, get_translated_strings, set_translated_string,
     add_resource, get_resource, get_resources, setup, shutdown_session, add_language, get_languages, get_language,
+    get_po_metadata, update_po_metadata,
 ]